ai-video-pipeline/pyproject.toml
[project]
name = "ai-video-pipeline"
version = "0.1.0"
description = "ai-video-pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "google-genai>=1.62.0",
    "groq>=1.0.0",
    "moviepy>=2.2.1",
    "openai-whisper>=20250625",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
